remove sdkv2 `cloudflare_ruleset`
